requirements.txt
mysql-connector-python==8.1.0
cryptography==41.0.1
python-dotenv==1.0.0
faker==19.3.0 